+++ packages/eas-build-job/src/common.ts
@@ -48,8 +48,9 @@ export type ArchiveSource =
       /**
        * Url that can be used to clone repository.
        * It should contain embedded credentials for private registries.
+       * May be omitted in remote jobs, which resolve sources through the API.
        */
-      repositoryUrl: string;
+      repositoryUrl?: string;
       /** A Git ref - points to a branch head, tag head or a branch name. */
       gitRef: string | null;
       /**
@@ -79,7 +80,7 @@ export const ArchiveSourceSchema = Joi.object<ArchiveSource>({
   .when(Joi.object({ type: ArchiveSourceType.GIT }).unknown(), {
     then: Joi.object({
       type: Joi.string().valid(ArchiveSourceType.GIT).required(),
-      repositoryUrl: Joi.string().required(),
+      repositoryUrl: Joi.string(),
       gitCommitHash: Joi.string().required(),
       gitRef: Joi.string().allow(null).required(),
     }),
@@ -94,7 +95,7 @@ export const ArchiveSourceSchema = Joi.object<ArchiveSource>({
 export const ArchiveSourceSchemaZ = z.discriminatedUnion('type', [
   z.object({
     type: z.literal(ArchiveSourceType.GIT),
-    repositoryUrl: z.string().url(),
+    repositoryUrl: z.string().url().optional(),
     gitRef: z.string().nullable(),
     gitCommitHash: z.string(),
   }),